==> code/Step 0 - Store path to scratch folder, home folder and shared folder.ipynb <==
In this project it was required to define different paths in the filesystem due to storage restrictions on `/home` path. 
- `/home` was used to store the code as well as few other files such as static files.
- `/scratch` was used to be able to store files temporarily that could be generated on demand.
- Additionally a `shared_data` space was available in scratch for members on the project to collaborate in, apart from their defined allocated home and scratch personal spaces.

**Note:** The below code works well on Linux machines. For users running Jupyter on Windows, the path definitions will need to be altered

import pickle

# Define base path where files will be stored
baseScratchPath = '/scratch/siads696s012w23_class_root/siads696s012w23_class/anilcm'
baseHomePath = '/home/anilcm'

# Define path to shared space in /scratch where the serialized files will reside
baseSharedPath = '/scratch/siads696s012w23_class_root/siads696s012w23_class/shared_data/team15/data'


pickledHomeScratchShared = (baseHomePath,baseScratchPath,baseSharedPath)

The below code is used to store the paths as pickle file in the **same level as code files** and unpickle them later through tuple unpacking.

with open("pickledHomeScratchShared.pickle", "wb") as f:
    pickle.dump(pickledHomeScratchShared, f)
